+++ demisto_sdk/commands/test_content/test_modeling_rule/tests/test_modeling_rule_test.py
@@ -227,6 +227,7 @@ class TestTheTestModelingRuleCommandSingleRule:
         from demisto_sdk.commands.test_content.xsiam_tools.test_data import TestData
 
         runner = CliRunner()
+        mocker.patch("time.sleep", return_value=None)
 
         # Create Test Data File
         pack.create_modeling_rule(DEFAULT_MODELING_RULE_NAME, rules=ONE_MODEL_RULE_TEXT)
@@ -249,7 +250,14 @@ class TestTheTestModelingRuleCommandSingleRule:
                     # Act
                     result = runner.invoke(
                         test_modeling_rule_cmd,
-                        [mrule_dir.as_posix(), "--non-interactive"],
+                        [
+                            mrule_dir.as_posix(),
+                            "--non-interactive",
+                            "--sleep_interval",
+                            "0",
+                            "--retry_attempts",
+                            "0",
+                        ],
                     )
                     # Assert
                     assert result.exit_code == 1
@@ -284,6 +292,7 @@ class TestTheTestModelingRuleCommandSingleRule:
         from demisto_sdk.commands.test_content.xsiam_tools.test_data import TestData
 
         runner = CliRunner()
+        mocker.patch("time.sleep", return_value=None)
 
         # Create Test Data File
         pack.create_modeling_rule(DEFAULT_MODELING_RULE_NAME, rules=ONE_MODEL_RULE_TEXT)
@@ -310,7 +319,14 @@ class TestTheTestModelingRuleCommandSingleRule:
                     # Act
                     result = runner.invoke(
                         test_modeling_rule_cmd,
-                        [mrule_dir.as_posix(), "--non-interactive"],
+                        [
+                            mrule_dir.as_posix(),
+                            "--non-interactive",
+                            "--sleep_interval",
+                            "0",
+                            "--retry_attempts",
+                            "0",
+                        ],
                     )
                     # Assert
                     assert result.exit_code == 1
@@ -340,26 +356,13 @@ class TestTheTestModelingRuleCommandSingleRule:
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
         monkeypatch.setenv("COLUMNS", "1000")
 
-        from functools import partial
-
-        from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
-            check_dataset_exists,
-        )
-
-        func_path = (
-            "demisto_sdk.commands.test_content.test_modeling_rule."
-            "test_modeling_rule.check_dataset_exists"
-        )
-        # override the default timeout to 1 second so only one iteration of the loop will be executed
-        check_dataset_exists_with_timeout = partial(check_dataset_exists, timeout=5)
-        monkeypatch.setattr(func_path, check_dataset_exists_with_timeout)
-
         from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
             app as test_modeling_rule_cmd,
         )
         from demisto_sdk.commands.test_content.xsiam_tools.test_data import TestData
 
         runner = CliRunner()
+        mocker.patch("time.sleep", return_value=None)
 
         # Create Test Data File
         pack.create_modeling_rule(DEFAULT_MODELING_RULE_NAME, rules=ONE_MODEL_RULE_TEXT)
@@ -393,7 +396,14 @@ class TestTheTestModelingRuleCommandSingleRule:
                     # Act
                     result = runner.invoke(
                         test_modeling_rule_cmd,
-                        [mrule_dir.as_posix(), "--non-interactive"],
+                        [
+                            mrule_dir.as_posix(),
+                            "--non-interactive",
+                            "--sleep_interval",
+                            "0",
+                            "--retry_attempts",
+                            "0",
+                        ],
                     )
                     # Assert
                     assert result.exit_code == 1
@@ -425,26 +435,13 @@ class TestTheTestModelingRuleCommandSingleRule:
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
         monkeypatch.setenv("COLUMNS", "1000")
 
-        from functools import partial
-
-        from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
-            check_dataset_exists,
-        )
-
-        func_path = (
-            "demisto_sdk.commands.test_content.test_modeling_rule."
-            "test_modeling_rule.check_dataset_exists"
-        )
-        # override the default timeout to 1 second so only one iteration of the loop will be executed
-        check_dataset_exists_with_timeout = partial(check_dataset_exists, timeout=5)
-        monkeypatch.setattr(func_path, check_dataset_exists_with_timeout)
-
         from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
             app as test_modeling_rule_cmd,
         )
         from demisto_sdk.commands.test_content.xsiam_tools.test_data import TestData
 
         runner = CliRunner()
+        mocker.patch("time.sleep", return_value=None)
 
         # Create Test Data File
         pack.create_modeling_rule(DEFAULT_MODELING_RULE_NAME, rules=ONE_MODEL_RULE_TEXT)
@@ -492,7 +489,14 @@ class TestTheTestModelingRuleCommandSingleRule:
                     # Act
                     result = runner.invoke(
                         test_modeling_rule_cmd,
-                        [mrule_dir.as_posix(), "--non-interactive"],
+                        [
+                            mrule_dir.as_posix(),
+                            "--non-interactive",
+                            "--sleep_interval",
+                            "0",
+                            "--retry_attempts",
+                            "0",
+                        ],
                     )
                     # Assert
                     assert result.exit_code == 1
@@ -524,26 +528,13 @@ class TestTheTestModelingRuleCommandSingleRule:
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
         monkeypatch.setenv("COLUMNS", "1000")
 
-        from functools import partial
-
-        from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
-            check_dataset_exists,
-        )
-
-        func_path = (
-            "demisto_sdk.commands.test_content.test_modeling_rule."
-            "test_modeling_rule.check_dataset_exists"
-        )
-        # override the default timeout to 1 second so only one iteration of the loop will be executed
-        check_dataset_exists_with_timeout = partial(check_dataset_exists, timeout=5)
-        monkeypatch.setattr(func_path, check_dataset_exists_with_timeout)
-
         from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
             app as test_modeling_rule_cmd,
         )
         from demisto_sdk.commands.test_content.xsiam_tools.test_data import TestData
 
         runner = CliRunner()
+        mocker.patch("time.sleep", return_value=None)
 
         # Create Test Data File
         pack.create_modeling_rule(DEFAULT_MODELING_RULE_NAME, rules=ONE_MODEL_RULE_TEXT)
@@ -596,7 +587,14 @@ class TestTheTestModelingRuleCommandSingleRule:
                     # Act
                     result = runner.invoke(
                         test_modeling_rule_cmd,
-                        [mrule_dir.as_posix(), "--non-interactive"],
+                        [
+                            mrule_dir.as_posix(),
+                            "--non-interactive",
+                            "--sleep_interval",
+                            "0",
+                            "--retry_attempts",
+                            "0",
+                        ],
                     )
                     # Assert
                     assert result.exit_code == 1
@@ -627,23 +625,13 @@ class TestTheTestModelingRuleCommandSingleRule:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
 
-        from functools import partial
-
-        from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
-            check_dataset_exists,
-        )
-
-        monkeypatch.setattr(
-            "demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule.check_dataset_exists",
-            partial(check_dataset_exists),
-        )
-
         from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
             app as test_modeling_rule_cmd,
         )
         from demisto_sdk.commands.test_content.xsiam_tools.test_data import TestData
 
         runner = CliRunner()
+        mocker.patch("time.sleep", return_value=None)
 
         # Create Test Data File
         pack.create_modeling_rule(DEFAULT_MODELING_RULE_NAME, rules=ONE_MODEL_RULE_TEXT)
@@ -692,24 +680,6 @@ class TestTheTestModelingRuleCommandSingleRule:
                                 "json": {
                                     "reply": {
                                         "status": "SUCCESS",
-                                        "results": {"data": []},
-                                    }
-                                },
-                                "status_code": 200,
-                            },
-                            {
-                                "json": {
-                                    "reply": {
-                                        "status": "SUCCESS",
-                                        "results": {"data": []},
-                                    }
-                                },
-                                "status_code": 200,
-                            },
-                            {
-                                "json": {
-                                    "reply": {
-                                        "status": "SUCCESS",
                                         "results": {
                                             "data": [
                                                 {
@@ -735,7 +705,14 @@ class TestTheTestModelingRuleCommandSingleRule:
                     # Act
                     result = runner.invoke(
                         test_modeling_rule_cmd,
-                        [mrule_dir.as_posix(), "--non-interactive"],
+                        [
+                            mrule_dir.as_posix(),
+                            "--non-interactive",
+                            "--sleep_interval",
+                            "0",
+                            "--retry_attempts",
+                            "0",
+                        ],
                     )
                     # Assert
                     assert result.exit_code == 0
@@ -768,20 +745,6 @@ class TestTheTestModelingRuleCommandSingleRule:
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         monkeypatch.setenv("COLUMNS", "1000")
 
-        from functools import partial
-
-        from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
-            check_dataset_exists,
-        )
-
-        func_path = (
-            "demisto_sdk.commands.test_content.test_modeling_rule."
-            "test_modeling_rule.check_dataset_exists"
-        )
-        # override the default timeout to 1 second so only one iteration of the loop will be executed
-        check_dataset_exists_with_timeout = partial(check_dataset_exists, timeout=5)
-        monkeypatch.setattr(func_path, check_dataset_exists_with_timeout)
-
         from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
             app as test_modeling_rule_cmd,
         )
@@ -791,6 +754,7 @@ class TestTheTestModelingRuleCommandSingleRule:
         monkeypatch.setenv("COLUMNS", "1000")
 
         runner = CliRunner()
+        mocker.patch("time.sleep", return_value=None)
 
         # Create Test Data File
         pack.create_modeling_rule(DEFAULT_MODELING_RULE_NAME, rules=ONE_MODEL_RULE_TEXT)
@@ -865,7 +829,14 @@ class TestTheTestModelingRuleCommandSingleRule:
                     # Act
                     result = runner.invoke(
                         test_modeling_rule_cmd,
-                        [mrule_dir.as_posix(), "--non-interactive"],
+                        [
+                            mrule_dir.as_posix(),
+                            "--non-interactive",
+                            "--sleep_interval",
+                            "0",
+                            "--retry_attempts",
+                            "0",
+                        ],
                     )
                     # Assert
                     assert result.exit_code == 0
@@ -899,28 +870,16 @@ class TestTheTestModelingRuleCommandSingleRule:
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         monkeypatch.setenv("COLUMNS", "1000")
 
-        from functools import partial
-
         from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
             app as test_modeling_rule_cmd,
         )
-        from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
-            check_dataset_exists,
-        )
         from demisto_sdk.commands.test_content.xsiam_tools.test_data import TestData
-
-        func_path = (
-            "demisto_sdk.commands.test_content.test_modeling_rule."
-            "test_modeling_rule.check_dataset_exists"
-        )
-        # override the default timeout to 1 second so only one iteration of the loop will be executed
-        check_dataset_exists_with_timeout = partial(check_dataset_exists, timeout=5)
-        monkeypatch.setattr(func_path, check_dataset_exists_with_timeout)
 
         # so the logged output when running the command will be printed with a width of 120 characters
         monkeypatch.setenv("COLUMNS", "1000")
 
         runner = CliRunner()
+        mocker.patch("time.sleep", return_value=None)
 
         # Create Test Data File
         pack.create_modeling_rule(DEFAULT_MODELING_RULE_NAME, rules=ONE_MODEL_RULE_TEXT)
@@ -1000,7 +959,14 @@ class TestTheTestModelingRuleCommandSingleRule:
                     # Act
                     result = runner.invoke(
                         test_modeling_rule_cmd,
-                        [mrule_dir.as_posix(), "--non-interactive"],
+                        [
+                            mrule_dir.as_posix(),
+                            "--non-interactive",
+                            "--sleep_interval",
+                            "0",
+                            "--retry_attempts",
+                            "0",
+                        ],
                     )
                     # Assert
                     assert result.exit_code == 0
@@ -1017,7 +983,7 @@ class TestTheTestModelingRuleCommandSingleRule:
             assert False, "No exception should be raised in this scenario."
 
     def test_the_test_modeling_rule_command_results_with_non_existent_ignored_validations(
-        self, pack
+        self, pack, mocker
     ):
         """
         Given:
@@ -1042,6 +1008,7 @@ class TestTheTestModelingRuleCommandSingleRule:
         from demisto_sdk.commands.test_content.xsiam_tools.test_data import TestData
 
         runner = CliRunner()
+        mocker.patch("time.sleep", return_value=None)
 
         # Create Test Data File
         pack.create_modeling_rule(DEFAULT_MODELING_RULE_NAME, rules=ONE_MODEL_RULE_TEXT)
@@ -1071,7 +1038,14 @@ class TestTheTestModelingRuleCommandSingleRule:
                 # Act
                 result = runner.invoke(
                     test_modeling_rule_cmd,
-                    [mrule_dir.as_posix(), "--non-interactive"],
+                    [
+                        mrule_dir.as_posix(),
+                        "--non-interactive",
+                        "--sleep_interval",
+                        "0",
+                        "--retry_attempts",
+                        "0",
+                    ],
                 )
                 # Assert
                 assert result.exit_code == 1
@@ -1105,26 +1079,13 @@ class TestTheTestModelingRuleCommandSingleRule:
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
         monkeypatch.setenv("COLUMNS", "1000")
 
-        from functools import partial
-
-        from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
-            check_dataset_exists,
-        )
-
-        func_path = (
-            "demisto_sdk.commands.test_content.test_modeling_rule."
-            "test_modeling_rule.check_dataset_exists"
-        )
-        # override the default timeout to 1 second so only one iteration of the loop will be executed
-        check_dataset_exists_with_timeout = partial(check_dataset_exists, timeout=5)
-        monkeypatch.setattr(func_path, check_dataset_exists_with_timeout)
-
         from demisto_sdk.commands.test_content.test_modeling_rule.test_modeling_rule import (
             app as test_modeling_rule_cmd,
         )
         from demisto_sdk.commands.test_content.xsiam_tools.test_data import TestData
 
         runner = CliRunner()
+        mocker.patch("time.sleep", return_value=None)
 
         # Create Test Data File
         pack.create_modeling_rule(DEFAULT_MODELING_RULE_NAME, rules=ONE_MODEL_RULE_TEXT)
@@ -1196,7 +1157,14 @@ class TestTheTestModelingRuleCommandSingleRule:
                     # Act
                     result = runner.invoke(
                         test_modeling_rule_cmd,
-                        [mrule_dir.as_posix(), "--non-interactive"],
+                        [
+                            mrule_dir.as_posix(),
+                            "--non-interactive",
+                            "--sleep_interval",
+                            "0",
+                            "--retry_attempts",
+                            "0",
+                        ],
                     )
                     # Assert
                     assert result.exit_code == 1
@@ -1240,6 +1208,7 @@ class TestTheTestModelingRuleCommandMultipleRules:
         monkeypatch.setenv("COLUMNS", "1000")
 
         runner = CliRunner()
+        mocker.patch("time.sleep", return_value=None)
 
         # Create Pack 1 with Modeling Rule
         pack_1 = repo.create_pack("Pack1")
@@ -1340,6 +1309,10 @@ class TestTheTestModelingRuleCommandMultipleRules:
                             mrule_dir_1.as_posix(),
                             mrule_dir_2.as_posix(),
                             "--non-interactive",
+                            "--sleep_interval",
+                            "0",
+                            "--retry_attempts",
+                            "0",
                         ],
                     )
                     # Assert
@@ -1389,6 +1362,7 @@ class TestTheTestModelingRuleCommandInteractive:
         monkeypatch.setenv("COLUMNS", "1000")
 
         runner = CliRunner()
+        mocker.patch("time.sleep", return_value=None)
 
         # Create Pack with Modeling Rule
         pack = repo.create_pack("Pack1")
@@ -1413,7 +1387,15 @@ class TestTheTestModelingRuleCommandInteractive:
                 mock_prompt.return_value = 2
                 # Act
                 result = runner.invoke(
-                    test_modeling_rule_cmd, [mrule_dir.as_posix(), "--interactive"]
+                    test_modeling_rule_cmd,
+                    [
+                        mrule_dir.as_posix(),
+                        "--interactive",
+                        "--sleep_interval",
+                        "0",
+                        "--retry_attempts",
+                        "0",
+                    ],
                 )
                 # Assert